+++ pac4j-oidc/src/test/java/org/pac4j/oidc/metadata/OidcOpMetadataResolverTest.java
@@ -1,14 +1,12 @@
 package org.pac4j.oidc.metadata;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import com.nimbusds.oauth2.sdk.auth.ClientAuthenticationMethod;
 import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Set;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
-import static org.junit.jupiter.api.Assertions.fail;
+
 import org.pac4j.core.exception.TechnicalException;
 import org.pac4j.oidc.config.OidcConfiguration;