+++ pac4j-saml/src/main/java/org/pac4j/saml/redirect/SAML2RedirectionActionBuilder.java
@@ -1,16 +1,13 @@
 package org.pac4j.saml.redirect;
 
+import lombok.RequiredArgsConstructor;
 import lombok.val;
 import org.opensaml.saml.common.xml.SAMLConstants;
-import org.opensaml.saml.saml2.core.AuthnRequest;
 import org.pac4j.core.context.CallContext;
 import org.pac4j.core.exception.http.RedirectionAction;
 import org.pac4j.core.redirect.RedirectionActionBuilder;
-import org.pac4j.core.util.CommonHelper;
 import org.pac4j.core.util.HttpActionHelper;
 import org.pac4j.saml.client.SAML2Client;
-import org.pac4j.saml.profile.api.SAML2ObjectBuilder;
-import org.pac4j.saml.sso.impl.SAML2AuthnRequestBuilder;
 
 import java.util.Optional;
 
@@ -20,23 +17,18 @@ import java.util.Optional;
  * @author Jerome Leleu
  * @since 2.0.0
  */
+@RequiredArgsConstructor
 public class SAML2RedirectionActionBuilder implements RedirectionActionBuilder {
 
     private final SAML2Client client;
-    protected SAML2ObjectBuilder<AuthnRequest> saml2ObjectBuilder;
-
-    public SAML2RedirectionActionBuilder(final SAML2Client client) {
-        CommonHelper.assertNotNull("client", client);
-        this.client = client;
-        this.saml2ObjectBuilder = new SAML2AuthnRequestBuilder();
-    }
 
     @Override
     public Optional<RedirectionAction> getRedirectionAction(final CallContext ctx) {
         val context = this.client.getContextProvider().buildContext(ctx, this.client);
         val relayState = this.client.getStateGenerator().generateValue(ctx);
 
-        val authnRequest = this.saml2ObjectBuilder.build(context);
+        val saml2ObjectBuilder = client.getConfiguration().getSamlAuthnRequestBuilder();
+        val authnRequest = saml2ObjectBuilder.build(context);
         this.client.getWebSsoMessageSender().sendMessage(context, authnRequest, relayState);
 
         val adapter = context.getProfileRequestContextOutboundMessageTransportResponse();