+++ pac4j-saml/src/main/java/org/pac4j/saml/sso/impl/SAML2AuthnRequestBuilder.java
@@ -1,11 +1,22 @@
 package org.pac4j.saml.sso.impl;
 
+import lombok.Setter;
+import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.apache.commons.lang3.StringUtils;
 import org.opensaml.core.xml.XMLObjectBuilderFactory;
 import org.opensaml.saml.common.SAMLObjectBuilder;
 import org.opensaml.saml.common.SAMLVersion;
-import org.opensaml.saml.saml2.core.*;
+import org.opensaml.saml.saml2.core.AuthnContextClassRef;
+import org.opensaml.saml.saml2.core.AuthnContextComparisonTypeEnumeration;
+import org.opensaml.saml.saml2.core.AuthnRequest;
+import org.opensaml.saml.saml2.core.Extensions;
+import org.opensaml.saml.saml2.core.IDPEntry;
+import org.opensaml.saml.saml2.core.IDPList;
+import org.opensaml.saml.saml2.core.Issuer;
+import org.opensaml.saml.saml2.core.NameID;
+import org.opensaml.saml.saml2.core.Scoping;
+import org.opensaml.saml.saml2.core.Subject;
 import org.opensaml.saml.saml2.core.impl.AuthnContextClassRefBuilder;
 import org.opensaml.saml.saml2.core.impl.NameIDPolicyBuilder;
 import org.opensaml.saml.saml2.core.impl.RequestedAuthnContextBuilder;
@@ -19,6 +30,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.time.Instant;
+import java.util.Objects;
 
 /**
  * Build a SAML2 Authn Request from the given {@link org.opensaml.messaging.context.MessageContext}.
@@ -27,14 +39,16 @@ import java.time.Instant;
  * @author Misagh Moayyed
  * @since 1.5.0
  */
+@SuppressWarnings("unchecked")
+@Slf4j
 public class SAML2AuthnRequestBuilder implements SAML2ObjectBuilder<AuthnRequest> {
     protected final Logger logger = LoggerFactory.getLogger(getClass());
 
+    @Setter
     private int issueInstantSkewSeconds = 0;
 
     private final XMLObjectBuilderFactory builderFactory = Configuration.getBuilderFactory();
 
-    /** {@inheritDoc} */
     @Override
     public AuthnRequest build(final SAML2MessageContext context) {
         val configContext = context.getConfigurationContext();
@@ -49,15 +63,6 @@ public class SAML2AuthnRequestBuilder implements SAML2ObjectBuilder<AuthnRequest
         return authnRequest;
     }
 
-    /**
-     * <p>buildAuthnRequest.</p>
-     *
-     * @param context a {@link SAML2MessageContext} object
-     * @param assertionConsumerService a {@link org.opensaml.saml.saml2.metadata.AssertionConsumerService} object
-     * @param ssoService a {@link org.opensaml.saml.saml2.metadata.SingleSignOnService} object
-     * @return a {@link AuthnRequest} object
-     */
-    @SuppressWarnings("unchecked")
     protected final AuthnRequest buildAuthnRequest(final SAML2MessageContext context,
                                                    final Endpoint assertionConsumerService,
                                                    final Endpoint ssoService) {
@@ -161,29 +166,34 @@ public class SAML2AuthnRequestBuilder implements SAML2ObjectBuilder<AuthnRequest
                 request.setScoping(scoping);
             }
         }
+        return buildAuthnRequestSubject(context, request);
+    }
+
+    protected AuthnRequest buildAuthnRequestSubject(final SAML2MessageContext context,
+                                                    final AuthnRequest request) {
+        val configContext = context.getConfigurationContext();
+        val nameIdValue = configContext.getSAML2Configuration().getAuthnRequestSubjectNameId();
+        val nameIdFormat = configContext.getSAML2Configuration().getAuthnRequestSubjectNameIdFormat();
+        if (StringUtils.isNotBlank(nameIdValue) && StringUtils.isNotBlank(nameIdFormat)) {
+            val subject = ((SAMLObjectBuilder<Subject>) Objects.requireNonNull(this.builderFactory
+                .getBuilder(Subject.DEFAULT_ELEMENT_NAME))).buildObject();
+            val nameId = ((SAMLObjectBuilder<NameID>) Objects.requireNonNull(this.builderFactory
+                .getBuilder(NameID.DEFAULT_ELEMENT_NAME))).buildObject();
+            LOGGER.debug("Setting AuthnRequest Subject NameID with value [{}] and format [{}]", nameIdValue, nameIdFormat);
+            nameId.setValue(nameIdValue);
+            nameId.setFormat(nameIdFormat);
+            subject.setNameID(nameId);
+            request.setSubject(subject);
+        }
         return request;
     }
 
-    /**
-     * <p>buildAuthnContextClassRef.</p>
-     *
-     * @param authnContextClassRef a {@link String} object
-     * @return a {@link AuthnContextClassRef} object
-     */
     protected AuthnContextClassRef buildAuthnContextClassRef(final String authnContextClassRef) {
         val classRef = new AuthnContextClassRefBuilder().buildObject();
         classRef.setURI(authnContextClassRef);
         return classRef;
     }
 
-    /**
-     * <p>getIssuer.</p>
-     *
-     * @param context a {@link SAML2MessageContext} object
-     * @param spEntityId a {@link String} object
-     * @return a {@link Issuer} object
-     */
-    @SuppressWarnings("unchecked")
     protected final Issuer getIssuer(final SAML2MessageContext context, final String spEntityId) {
         val configContext = context.getConfigurationContext();
 
@@ -201,12 +211,6 @@ public class SAML2AuthnRequestBuilder implements SAML2ObjectBuilder<AuthnRequest
         return issuer;
     }
 
-    /**
-     * <p>getComparisonTypeEnumFromString.</p>
-     *
-     * @param comparisonType a {@link String} object
-     * @return a {@link AuthnContextComparisonTypeEnumeration} object
-     */
     protected AuthnContextComparisonTypeEnumeration getComparisonTypeEnumFromString(final String comparisonType) {
         if ("exact".equalsIgnoreCase(comparisonType)) {
             return AuthnContextComparisonTypeEnumeration.EXACT;
@@ -221,14 +225,5 @@ public class SAML2AuthnRequestBuilder implements SAML2ObjectBuilder<AuthnRequest
             return AuthnContextComparisonTypeEnumeration.BETTER;
         }
         return null;
-    }
-
-    /**
-     * <p>Setter for the field <code>issueInstantSkewSeconds</code>.</p>
-     *
-     * @param issueInstantSkewSeconds a int
-     */
-    public void setIssueInstantSkewSeconds(final int issueInstantSkewSeconds) {
-        this.issueInstantSkewSeconds = issueInstantSkewSeconds;
     }
 }